Particle in a river
This is my jupyter note book for my particle in a river. The conditions are set up so that the particle gets pushed enough by the river to complete 1 revolution and go back into the river. Increasing the drag coefficient results in the particle not completing a revolution and stopping verifying my model.